+++ src/shared/tdoc-commons/services/tdoc-sql-mytbdb.adapter.ts
@@ -307,6 +307,12 @@ export class TourDocSqlMytbDbAdapter extends GenericSqlAdapter<TourDocRecord, To
             return this.actionTagODAdapter.executeActionTagObjectsKey(table, id, <ObjectsKeyActionTagForm> actionTagForm, opts);
         } else if ((table === 'image' || table === 'video') && actionTagForm.type === 'tag' && actionTagForm.key.startsWith('persRate_')) {
             return this.actionTagRateAdapter.executeActionTagRate(table, id, <RateActionTagForm> actionTagForm, opts);
+        } else if ((table === 'track') && actionTagForm.type === 'tag' && actionTagForm.key.startsWith('trackmedia_persRate')) {
+            return this.actionTagRateAdapter.executeActionTagRate('trackimages', id, <RateActionTagForm> actionTagForm, opts).then(
+                () => {
+                    return this.actionTagRateAdapter.executeActionTagRate('trackvideos', id, <RateActionTagForm> actionTagForm, opts);
+                }
+            );
         } else if (actionTagForm.type === 'tag' && actionTagForm.key.startsWith('blocked')) {
             return this.actionTagBlockAdapter.executeActionTagBlock(table, id, actionTagForm, opts);
         } else if (actionTagForm.type === 'assign' && actionTagForm.key.startsWith('assign')) {